# Issue 266: refactor supply chain structure input datasets and various speed ups