--- a/celavi/compute_locations.py
+++ b/celavi/compute_locations.py
@@ -987,10 +987,12 @@ class ComputeLocations:
 
         # exclude Hawaii, Guam, Puerto Rico, and Alaska
         # (only have road network data for the contiguous United States)
+        # DC has no landfills in the landfill MOP data
         locations = locations[locations.region_id_2 != 'GU']
         locations = locations[locations.region_id_2 != 'HI']
         locations = locations[locations.region_id_2 != 'PR']
         locations = locations[locations.region_id_2 != 'AK']
+        locations = locations[locations.region_id_2 != 'DC']
         
         locations = locations[locations.region_id_3 != 'Nantucket']