We've been having consistent issues with defining and verifying the network structure, and having to manage half a dozen different input files with this information isn't helping. The `CostGraph` instantiation process was taking 10-15 minutes which is excessive for the network size we're working with. `Router` ditto, was finding far more routes than were actually needed in the network.

This PR and the associated [data PR](https://github.nrel.gov/rhanes/celavi-data/pull/99) (national scale only for the moment) accomplish:

- Replaces most of the network structure input files with a single, relatively human readable csv file called network-structure.csv - see attached.
- Works from the network-structure file and the locations_computed data structure to generate a dataframe, "network_edges" (attached), that defines all the edges and metadata required to instantiate `CostGraph`, except for the distances, route_ids (obtained via `Router.get_all_routes`), and numerical costs (calculated during instantiation). network_edges is generated in scenario.py.
- Refactors `Router.get_all_routes` to run off the network_edges data structure instead of the locations_computed file. This cuts down the number of calls to the router and provides a handy way to identify edges with no distances/routes (ie colocated nodes) and edges that should not be included in `CostGraph` (ie in-state connections).
    - I sped up all the non-route-finding operations in `Router.get_all_routes`, and current speed is right around 100 routes per minute. For the glass case study that works out to a little under 90 minutes, although I'm seeing substantial variation in that time on my machine. I implemented some timing during development and confirmed the holdup is indeed the actual route finding, and so this is about as good as it gets. I still recommend using pre-computed routes when debugging locally.
- The routes_computed file has the same name but looks substantially different: it's network_edges with additional columns for vkmt and route_id. As a bonus, this file is now ~5 MB and can live on Git instead of Kestrel.
- Refactors `CostGraph.build_supplychain_graph` to be a self-contained method - no more sub-methods that define facilities/nodes/edges - that first processes the network_edges file to generate a list of edges in a networkx-friendly format, then creates the network from the list of edges in a single step, followed by the addition of the facility_id node attribute (required for pathfinding logic). After refactorization, CostGraph instantiation is essentially instant even with the full national data set (glass study).
- Incidentally, because of the way `Router.get_all_routes` now executes, it's simpler to save the routes-and-distances-by-county data in case we ever want to use that to disaggregate transpo emissions by county traversed. This file is vkmt-county-all.csv (available in the data repo).
- In testing the above new features, several bugs and logic errors were identified and fixed, **including one bug that was likely preventing landfilling from being chosen as an EOL pathway**. See commit [a461d2e](https://github.com/NREL/celavi/pull/267/commits/a461d2e2254ad8471e7853d25a50684a2f5718cb).
- Other bugs identified and fixes implemented during testing: 
    - Some nodes lost all downstream connections to landfills due to the vkmt_max logic; Router now checks for this and adds in the closest landfill to only these nodes regardless of vkmt_max.
    - Washington, DC power plants and buildings had no downstream landfills because there are no DC landfills in the LMOP dataset; compute_locations now filters out DC facilities in addition to Alaska, Hawaii, etc. We could alternatively add in fake landfills in DC if we want to keep the DC power plants and buildings.

Testing on the national data has completed. Output and diagnostic results are below.

The latest CostGraph file is also attached for review.

[network-structure.csv](https://github.com/user-attachments/files/19411841/network-structure.csv)

[network_edges.csv](https://github.com/user-attachments/files/19437185/network_edges.csv)

[netw.csv](https://github.com/user-attachments/files/19437188/netw.csv)

Output from completed national run w/out computing routes:

```
(celavisandbox3) C:\Users\rhanes\GitHub\celavi>python -m celavi --data C:\Users\rhanes\GitHub\celavi-data --casestudy casestudy.yaml --scenario scenario.yaml
Preprocessing starting at 0 s
validating PVTechUnitLocations
validated PVTechUnitLocations
no missing data values in celavi.data_manager.PVTechUnitLocations.eia_id
no missing data values in celavi.data_manager.PVTechUnitLocations.p_year
validating CommWindowTechUnitLocations
validated CommWindowTechUnitLocations
no missing data values in celavi.data_manager.CommWindowTechUnitLocations.year
validating LandfillLocations
validated LandfillLocations
279 of 2637 data values in celavi.data_manager.LandfillLocations.Landfill Closure Year were backfilled as -1
validating OtherFacilityLocations
validated OtherFacilityLocations
validating StandardScenarios
validated StandardScenarios
no missing data values in celavi.data_manager.StandardScenarios.wind_onshore_MW
no missing data values in celavi.data_manager.StandardScenarios.upv_MW
validating PVTechUnitChars
validated PVTechUnitChars
no missing data values in celavi.data_manager.PVTechUnitChars.year
no missing data values in celavi.data_manager.PVTechUnitChars.MWdc_per_m2
no missing data values in celavi.data_manager.PVTechUnitChars.MWdc_per_module
no missing data values in celavi.data_manager.PVTechUnitChars.glass_metrictonne_per_module
validating StateCentroids
validated StateCentroids
no missing data values in celavi.data_manager.StateCentroids.lat
no missing data values in celavi.data_manager.StateCentroids.long
Run routes completed at 1 s
CostGraph instantiation started at 1 s
CostGraph: Adding nodes and edges
CostGraph: Adding nodes and edges took 0.0 minutes
CostGraph: Calculating edge costs
CostGraph: Calculating edge costs took 0.0 minutes
CostGraph: Adjusting all costs for 2000 upwards by $4417.0
CostGraph: Instantiation took 0.0 minutes
CostGraph instantiated at 1 s
Context initialized at 2 s
Instantiating 2495.0 components at 2 s
100 components instantiated
200 components instantiated
300 components instantiated
400 components instantiated
500 components instantiated
600 components instantiated
700 components instantiated
800 components instantiated
900 components instantiated
1000 components instantiated
1100 components instantiated
1200 components instantiated
1300 components instantiated
1400 components instantiated
1500 components instantiated
1600 components instantiated
1700 components instantiated
1800 components instantiated
1900 components instantiated
2000 components instantiated
2100 components instantiated
2200 components instantiated
2300 components instantiated
2400 components instantiated
Instantiating components took 10.11 minutes
Beginning discrete event simulation at 609 s
CostGraph.update_costs: Adjusting all costs for 2001.0 upwards by $4417.0
CostGraph.update_costs: Adjusting all costs for 2002.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2002.0 after 9.1 s
CostGraph.update_costs: Adjusting all costs for 2003.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2003.0 after 7.5 s
CostGraph.update_costs: Adjusting all costs for 2004.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2004.0 after 6.1 s
CostGraph.update_costs: Adjusting all costs for 2005.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2005.0 after 10.3 s
CostGraph.update_costs: Adjusting all costs for 2006.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2006.0 after 10.1 s
CostGraph.update_costs: Adjusting all costs for 2007.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2007.0 after 6.2 s
CostGraph.update_costs: Adjusting all costs for 2008.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2008.0 after 5.7 s
CostGraph.update_costs: Adjusting all costs for 2009.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2009.0 after 7.3 s
CostGraph.update_costs: Adjusting all costs for 2010.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2010.0 after 9.4 s
CostGraph.update_costs: Adjusting all costs for 2011.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2011.0 after 8.7 s
CostGraph.update_costs: Adjusting all costs for 2012.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2012.0 after 8.5 s
CostGraph.update_costs: Adjusting all costs for 2013.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2013.0 after 9.9 s
CostGraph.update_costs: Adjusting all costs for 2014.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2014.0 after 7.4 s
CostGraph.update_costs: Adjusting all costs for 2015.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2015.0 after 6.5 s
CostGraph.update_costs: Adjusting all costs for 2016.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2016.0 after 9.8 s
CostGraph.update_costs: Adjusting all costs for 2017.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2017.0 after 13.9 s
CostGraph.update_costs: Adjusting all costs for 2018.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2018.0 after 12.2 s
CostGraph.update_costs: Adjusting all costs for 2019.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2019.0 after 16.2 s
CostGraph.update_costs: Adjusting all costs for 2020.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2020.0 after 13.6 s
CostGraph.update_costs: Adjusting all costs for 2021.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2021.0 after 13.5 s
CostGraph.update_costs: Adjusting all costs for 2022.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2022.0 after 7.6 s
CostGraph.update_costs: Adjusting all costs for 2023.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2023.0 after 19.0 s
CostGraph.update_costs: Adjusting all costs for 2024.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2024.0 after 12.2 s
CostGraph.update_costs: Adjusting all costs for 2025.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2025.0 after 11.9 s
CostGraph.update_costs: Adjusting all costs for 2026.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2026.0 after 9.6 s
CostGraph.update_costs: Adjusting all costs for 2027.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2027.0 after 21.5 s
CostGraph.update_costs: Adjusting all costs for 2028.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2028.0 after 24.9 s
CostGraph.update_costs: Adjusting all costs for 2029.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2029.0 after 32.6 s
CostGraph.update_costs: Adjusting all costs for 2030.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2030.0 after 37.0 s
CostGraph.update_costs: Adjusting all costs for 2031.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2031.0 after 1813.1 s
CostGraph.update_costs: Adjusting all costs for 2032.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2032.0 after 959.5 s
CostGraph.update_costs: Adjusting all costs for 2033.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2033.0 after 1195.4 s
CostGraph.update_costs: Adjusting all costs for 2034.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2034.0 after 879.8 s
CostGraph.update_costs: Adjusting all costs for 2035.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2035.0 after 705.0 s
CostGraph.update_costs: Adjusting all costs for 2036.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2036.0 after 932.8 s
CostGraph.update_costs: Adjusting all costs for 2037.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2037.0 after 899.4 s
CostGraph.update_costs: Adjusting all costs for 2038.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2038.0 after 939.3 s
CostGraph.update_costs: Adjusting all costs for 2039.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2039.0 after 756.4 s
CostGraph.update_costs: Adjusting all costs for 2040.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2040.0 after 960.5 s
CostGraph.update_costs: Adjusting all costs for 2041.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2041.0 after 826.9 s
CostGraph.update_costs: Adjusting all costs for 2042.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2042.0 after 742.7 s
CostGraph.update_costs: Adjusting all costs for 2043.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2043.0 after 730.4 s
CostGraph.update_costs: Adjusting all costs for 2044.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2044.0 after 665.6 s
CostGraph.update_costs: Adjusting all costs for 2045.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2045.0 after 633.0 s
CostGraph.update_costs: Adjusting all costs for 2046.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2046.0 after 688.3 s
CostGraph.update_costs: Adjusting all costs for 2047.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2047.0 after 779.6 s
CostGraph.update_costs: Adjusting all costs for 2048.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2048.0 after 625.9 s
CostGraph.update_costs: Adjusting all costs for 2049.0 upwards by $4417.0
CostGraph.update_costs: Costs updated for 2049.0 after 794.2 s
Creating diagnostic visualizations for run 0 at 17520 s
Traceback (most recent call last):
  File "C:\Users\rhanes\AppData\Local\anaconda3\envs\celavisandbox3\lib\runpy.py", line 194, in _run_module_as_main
    return _run_code(code, main_globals, None,
  File "C:\Users\rhanes\AppData\Local\anaconda3\envs\celavisandbox3\lib\runpy.py", line 87, in _run_code
    exec(code, run_globals)
  File "C:\Users\rhanes\GitHub\celavi\celavi\__main__.py", line 25, in <module>
    Scenario(parser=PARSER)
  File "C:\Users\rhanes\GitHub\celavi\celavi\scenario.py", line 117, in __init__
    self.postprocess()
  File "C:\Users\rhanes\GitHub\celavi\celavi\scenario.py", line 613, in postprocess
    lcia_df = pd.read_csv(self.files["lcia_to_des"], names=lcia_names)
  File "C:\Users\rhanes\AppData\Local\anaconda3\envs\celavisandbox3\lib\site-packages\pandas\io\parsers\readers.py", line 912, in read_csv
    return _read(filepath_or_buffer, kwds)
  File "C:\Users\rhanes\AppData\Local\anaconda3\envs\celavisandbox3\lib\site-packages\pandas\io\parsers\readers.py", line 577, in _read
    parser = TextFileReader(filepath_or_buffer, **kwds)
  File "C:\Users\rhanes\AppData\Local\anaconda3\envs\celavisandbox3\lib\site-packages\pandas\io\parsers\readers.py", line 1407, in __init__
    self._engine = self._make_engine(f, self.engine)
  File "C:\Users\rhanes\AppData\Local\anaconda3\envs\celavisandbox3\lib\site-packages\pandas\io\parsers\readers.py", line 1661, in _make_engine
    self.handles = get_handle(
  File "C:\Users\rhanes\AppData\Local\anaconda3\envs\celavisandbox3\lib\site-packages\pandas\io\common.py", line 859, in get_handle
    handle = open(
FileNotFoundError: [Errno 2] No such file or directory: 'C:\\Users\\rhanes\\GitHub\\celavi-data\\generated/final_lcia_results_to_des.csv'
```

Diagnostic results, national scale, following commits [f2d76b9](https://github.com/NREL/celavi/pull/267/commits/f2d76b92bb341639f64be4f5188d1f249ffca8f2) and [b8cc2a6](https://github.com/NREL/celavi/pull/267/commits/b8cc2a61b3fb835d2f2d20a7a9de349743125a2b) that fix the lifespan/timeout issue displayed in the above results:

These are from a national run only until 2021, in the interests of time - landfilling does not appear because no components make it all the way to landfilling during this time span.

![component_counts_0](https://github.com/user-attachments/assets/f3ebae84-0c94-444a-a8dd-53ddf5a3dc5d)

[count_cumulative_histories.csv](https://github.com/user-attachments/files/19450136/count_cumulative_histories.csv)

![material_mass_0](https://github.com/user-attachments/assets/8c9d4a07-4ffe-41e6-aba4-42c16caf209e)


Diagnostics from a 2000 - 2050 run on issue-98tiny (tiny-circfutures data):

![component_counts_0](https://github.com/user-attachments/assets/53a0cd7a-ab38-4a69-ade5-d5995a0a5631)

![material_mass_0](https://github.com/user-attachments/assets/ba1156d8-3314-43fb-b9e2-6751cc794885)

[count_cumulative_histories.csv](https://github.com/user-attachments/files/19452363/count_cumulative_histories.csv)